Move setParent to noteService
