+++ src/domain/index.ts
@@ -29,7 +29,7 @@ const repositories = initRepositories(apiUrl, eventBus);
 const noteService = new NoteService(repositories.note);
 const editorToolsService = new EditorToolsService(repositories.editorTools);
 const noteListService = new NoteListService(repositories.note);
-const noteSettingsService = new NoteSettingsService(repositories.noteSettings, repositories.note);
+const noteSettingsService = new NoteSettingsService(repositories.noteSettings);
 const authService = new AuthService(eventBus, repositories.auth);
 const userService = new UserService(eventBus, repositories.user);
 const marketplaceService = new MarketplaceService(repositories.marketplace);
