+++ src/domain/noteSettings.service.ts
@@ -1,5 +1,4 @@
 import type NoteSettingsRepository from '@/domain/noteSettings.repository.interface';
-import type NoteRepository from '@/domain/note.repository.interface';
 import type NoteSettings from '@/domain/entities/NoteSettings';
 import type { NoteId } from './entities/Note';
 import NotFoundError from './entities/errors/NotFound';
@@ -16,18 +15,11 @@ export default class NoteSettingsService {
   private readonly noteSettingsRepository: NoteSettingsRepository;
 
   /**
-   * Note repository
-   */
-  private readonly noteRepository: NoteRepository;
-
-  /**
    * Note Service constructor
    * @param noteSettingsRepository - Note settings repository instance
-   * @param noteRepository - Note repository instance
    */
-  constructor(noteSettingsRepository: NoteSettingsRepository, noteRepository: NoteRepository) {
+  constructor(noteSettingsRepository: NoteSettingsRepository) {
     this.noteSettingsRepository = noteSettingsRepository;
-    this.noteRepository = noteRepository;
   }
 
   /**
@@ -109,36 +101,5 @@ export default class NoteSettingsService {
    */
   public async deleteNote(id: NoteId): Promise<void> {
     return await this.noteSettingsRepository.deleteNote(id);
-  }
-
-  /**
-   * Set new parent for the note
-   * @param id - Note id
-   * @param parentURL - link to the new parent note
-   */
-  public async setParent(id: NoteId, parentURL: string): Promise<void> {
-    // Regex matches any characters between '/note/' and the next slash or end of string
-    const regex = /\/note\/([^/]+)/;
-
-    const matches = parentURL.match(regex);
-
-    if (matches === null) {
-      throw new Error('Invalid parent URL');
-    }
-
-    // Extracts the ID from the URL. The ID is matches[1] as matches[0] is the full match
-    const parentId = matches[1];
-
-    const noteIdPattern = /^[a-zA-Z0-9-_]{10}$/;
-
-    if (!noteIdPattern.test(parentId)) {
-      throw new Error('Invalid parent ID');
-    }
-
-    const isUpdated = await this.noteRepository.setParent(id, parentId);
-
-    if (!isUpdated) {
-      throw new Error('Parent was not updated');
-    }
   }
 }
